some more updates
